src/{reflect,runtime}: break circular dependency from runtime to reflect
Use go:linkname, move code around.
