+++ src/runtime/hashmap.go
@@ -6,7 +6,6 @@ package runtime
 //     https://golang.org/src/runtime/map.go
 
 import (
-	"reflect"
 	"tinygo"
 	"unsafe"
 )
@@ -534,89 +533,8 @@ func hashmapStringDelete(m *hashmap, key string) {
 
 // Hashmap with interface keys (for everything else).
 
-// This is a method that is intentionally unexported in the reflect package. It
-// is identical to the Interface() method call, except it doesn't check whether
-// a field is exported and thus allows circumventing the type system.
-// The hash function needs it as it also needs to hash unexported struct fields.
-//
-//go:linkname valueInterfaceUnsafe reflect.valueInterfaceUnsafe
-func valueInterfaceUnsafe(v reflect.Value) interface{}
-
-func hashmapFloat32Hash(ptr unsafe.Pointer, seed uintptr) uint32 {
-	f := *(*uint32)(ptr)
-	if f == 0x80000000 {
-		// convert -0 to 0 for hashing
-		f = 0
-	}
-	return hash32(unsafe.Pointer(&f), 4, seed)
-}
-
-func hashmapFloat64Hash(ptr unsafe.Pointer, seed uintptr) uint32 {
-	f := *(*uint64)(ptr)
-	if f == 0x8000000000000000 {
-		// convert -0 to 0 for hashing
-		f = 0
-	}
-	return hash32(unsafe.Pointer(&f), 8, seed)
-}
-
-func hashmapInterfaceHash(itf interface{}, seed uintptr) uint32 {
-	x := reflect.ValueOf(itf)
-	if x.RawType() == nil {
-		return 0 // nil interface
-	}
-
-	value := (*_interface)(unsafe.Pointer(&itf)).value
-	ptr := value
-	if x.RawType().Size() <= unsafe.Sizeof(uintptr(0)) {
-		// Value fits in pointer, so it's directly stored in the pointer.
-		ptr = unsafe.Pointer(&value)
-	}
-
-	switch x.RawType().Kind() {
-	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-		return hash32(ptr, x.RawType().Size(), seed)
-	case reflect.Bool, reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64, reflect.Uintptr:
-		return hash32(ptr, x.RawType().Size(), seed)
-	case reflect.Float32:
-		// It should be possible to just has the contents. However, NaN != NaN
-		// so if you're using lots of NaNs as map keys (you shouldn't) then hash
-		// time may become exponential. To fix that, it would be better to
-		// return a random number instead:
-		// https://research.swtch.com/randhash
-		return hashmapFloat32Hash(ptr, seed)
-	case reflect.Float64:
-		return hashmapFloat64Hash(ptr, seed)
-	case reflect.Complex64:
-		rptr, iptr := ptr, unsafe.Add(ptr, 4)
-		return hashmapFloat32Hash(rptr, seed) ^ hashmapFloat32Hash(iptr, seed)
-	case reflect.Complex128:
-		rptr, iptr := ptr, unsafe.Add(ptr, 8)
-		return hashmapFloat64Hash(rptr, seed) ^ hashmapFloat64Hash(iptr, seed)
-	case reflect.String:
-		return hashmapStringHash(x.String(), seed)
-	case reflect.Chan, reflect.Ptr, reflect.UnsafePointer:
-		// It might seem better to just return the pointer, but that won't
-		// result in an evenly distributed hashmap. Instead, hash the pointer
-		// like most other types.
-		return hash32(ptr, x.RawType().Size(), seed)
-	case reflect.Array:
-		var hash uint32
-		for i := 0; i < x.Len(); i++ {
-			hash ^= hashmapInterfaceHash(valueInterfaceUnsafe(x.Index(i)), seed)
-		}
-		return hash
-	case reflect.Struct:
-		var hash uint32
-		for i := 0; i < x.NumField(); i++ {
-			hash ^= hashmapInterfaceHash(valueInterfaceUnsafe(x.Field(i)), seed)
-		}
-		return hash
-	default:
-		runtimePanic("comparing un-comparable type")
-		return 0 // unreachable
-	}
-}
+//go:linkname hashmapInterfaceHash reflect.hashmapInterfaceHash
+func hashmapInterfaceHash(itf interface{}, seed uintptr) uint32
 
 func hashmapInterfacePtrHash(iptr unsafe.Pointer, size uintptr, seed uintptr) uint32 {
 	_i := *(*interface{})(iptr)
